refactor NavbarRoutes && ProductDetailsRoutes in Routes directory
