--- src/Pages/Coupons/GetCoupon.js
@@ -1,5 +1,0 @@
-const GetCoupon = () => {
-  return <h1 className="mt-40">Get Coupons</h1>;
-};
-
-export default GetCoupon;
