--- src/Pages/OfferAndDiscount/OffersAndDiscounts.js
@@ -1,5 +1,0 @@
-const OffersAndDiscounts = () => {
-  return <h1 className="mt-40">Offers and Discounts</h1>;
-};
-
-export default OffersAndDiscounts;
