--- src/Pages/PrivacyPolicy/PrivacyPolicy.js
@@ -1,5 +1,0 @@
-const PrivacyPolicy = () => {
-  return <h1 className="mt-40">Privacy Policy</h1>;
-};
-
-export default PrivacyPolicy;
