--- src/Pages/RubberToys/RubberToys.js
@@ -1,9 +1,0 @@
-const RubberToys = () => {
-  return (
-    <div>
-      <h1 className="mt-40">Rubber Toys</h1>
-    </div>
-  );
-};
-
-export default RubberToys;
