--- src/Pages/Terms&Condition/TermsAndCondition.js
@@ -1,5 +1,0 @@
-const TermsAndCondition = () => {
-  return <h1 className="mt-40">Terms and Condition</h1>;
-};
-
-export default TermsAndCondition;
