--- src/Pages/WomanCloth/WomanCloth.js
@@ -1,5 +1,0 @@
-const WomanCloth = () => {
-  return <h1 className="mt-40">Woman Cloth</h1>;
-};
-
-export default WomanCloth;
