+++ src/Routes/NavbarRoutes.js
@@ -1,13 +1,13 @@
-import Blog from "../../Pages/Blog/Blog";
-import Contact from "../../Pages/Contact/Contact";
-import Shop from "../../Pages/Shop/Shop";
-import Latest from "../../Pages/Latest/Latest";
-import Home from "../../Pages/Home/Home";
-import About from "../../Pages/About/About";
-import CartPage from "../../Pages/CartPage/cart-page";
-import AuthPage from "../../Pages/Auth/Auth";
-import ProductDetails from "../../Pages/ProductDetails/ProductDetails";
-import CategoryPage from "../../Pages/Category/Category";
+import Blog from "../Pages/Blog/Blog";
+import Contact from "../Pages/Contact/Contact";
+import Shop from "../Pages/Shop/Shop";
+import Latest from "../Pages/Latest/Latest";
+import Home from "../Pages/Home/Home";
+import About from "../Pages/About/About";
+import CartPage from "../Pages/CartPage/cart-page";
+import AuthPage from "../Pages/Auth/Auth";
+import ProductDetails from "../Pages/ProductDetails/ProductDetails";
+import CategoryPage from "../Pages/Category/Category";
 
 export const NavbarRoutes = [
   {
