+++ src/Routes/routes.js
@@ -1,5 +1,5 @@
-import { NavbarRoutes } from "./NavbarRoutes/NavbarRoutes"
-import { ProductDetailsRoute } from "./ProductDetails/productsDetailsRoute";
+import { NavbarRoutes } from "./NavbarRoutes"
+import { ProductDetailsRoute } from "./productsDetailsRoute";
 import { CategoryRoute } from "./CategoryRoute";
 
 export const routes = [
